refactor (LadderMatchResult) : step4 - 게임결과 생성은 LadderMatchResult 클래스 내에서 생성할 수 있도록 수정 * 불필요한 getter를 피하기 위해 기능의 위치 변경
