+++ src/main/java/ladder/service/LadderService.java
@@ -2,10 +2,8 @@ package ladder.service;
 
 import ladder.controller.response.ResultDto;
 import ladder.domain.*;
-import ladder.service.type.GameResult;
+import ladder.domain.GameResult;
 import ladder.view.exception.InvalidInputToGetResult;
-
-import java.util.List;
 
 public class LadderService {
     private static final String GET_ALL_COMMENT = "all";
@@ -28,15 +26,7 @@ public class LadderService {
 
     public static GameResult getLadderGameResult(Persons persons, Ladder ladder, Rewards rewards) {
         LadderMatchResult ladderMatchResult = ladder.play(persons.getCount());
-        List<Person> personList = persons.getPersons();
 
-        int countOfPerson = personList.size();
-        GameResult gameResult = GameResult.getInstance();
-        for (int i = 0; i < countOfPerson; i++) {
-            int resultIndex = ladderMatchResult.get(i);
-            gameResult.put(Person.nameOf(personList.get(i)), rewards.getValue(resultIndex));
-        }
-
-        return gameResult;
+        return ladderMatchResult.map(persons, rewards);
     }
 }
