+++ src/test/java/ladder/service/LadderServiceTest.java
@@ -1,7 +1,7 @@
 package ladder.service;
 
 import ladder.controller.response.ResultDto;
-import ladder.service.type.GameResult;
+import ladder.domain.GameResult;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
